:truck: Rename module (#GI-28)
