+++ src/parsers/__init__.py
@@ -1,1 +1,1 @@
-from .connect_authorize_form_data_parser import *
+from .connect_authorize_form_data import *
